requirements.txt
streamlit==1.31.1
googlemaps==4.10.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
openai==1.12.0
geopy==2.4.1
polyline==2.0.0
flask==2.0.1
beautifulsoup4==4.12.2
pydeck==0.9.1